rename: RecordView, SettingView 파일 이동
