+++ lib/views/record_view.dart
@@ -4,7 +4,7 @@
 import 'package:bibletree/models/verse.dart';
 import 'package:bibletree/statics/app_statics.dart';
 import 'package:bibletree/viewModels/record_view_model.dart';
-import 'package:bibletree/views/record/text_input_view.dart';
+import 'package:bibletree/widgets/text_input_view.dart';
 import 'package:bibletree/widgets/verse_card.dart';
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
